refactor : Test 구조 리팩터링2
