+++ src/test/kotlin/com/flab/ticketing/order/repository/OrderRepositoryTest.kt
@@ -1,9 +1,9 @@
 package com.flab.ticketing.order.repository
 
-import com.flab.ticketing.testutils.OrderTestDataGenerator
-import com.flab.ticketing.testutils.PerformanceTestDataGenerator
+import com.flab.ticketing.testutils.generator.OrderTestDataGenerator
+import com.flab.ticketing.testutils.generator.PerformanceTestDataGenerator
 import com.flab.ticketing.testutils.RepositoryTest
-import com.flab.ticketing.testutils.UserTestDataGenerator
+import com.flab.ticketing.testutils.generator.UserTestDataGenerator
 import com.flab.ticketing.common.dto.service.CursorInfoDto
 import com.flab.ticketing.common.exception.BadRequestException
 import com.flab.ticketing.order.dto.request.OrderSearchConditions
