+++ src/test/kotlin/com/flab/ticketing/performance/entity/PerformanceDateTimeTest.kt
@@ -1,6 +1,6 @@
 package com.flab.ticketing.performance.entity
 
-import com.flab.ticketing.testutils.PerformanceTestDataGenerator
+import com.flab.ticketing.testutils.generator.PerformanceTestDataGenerator
 import com.flab.ticketing.testutils.UnitTest
 import com.flab.ticketing.common.exception.BusinessIllegalStateException
 import com.flab.ticketing.performance.exception.PerformanceErrorInfos
