+++ src/test/kotlin/com/flab/ticketing/performance/entity/PerformancePlaceSeatTest.kt
@@ -1,6 +1,6 @@
 package com.flab.ticketing.performance.entity
 
-import com.flab.ticketing.testutils.PerformanceTestDataGenerator
+import com.flab.ticketing.testutils.generator.PerformanceTestDataGenerator
 import com.flab.ticketing.testutils.UnitTest
 import com.flab.ticketing.common.entity.Region
 import com.flab.ticketing.common.exception.InvalidValueException
