+++ src/test/kotlin/com/flab/ticketing/testutils/generator/PerformanceTestDataGenerator.kt
@@ -1,4 +1,4 @@
-package com.flab.ticketing.testutils
+package com.flab.ticketing.testutils.generator
 
 import com.flab.ticketing.common.entity.Region
 import com.flab.ticketing.performance.entity.Performance
